test 2

+++ src/main/java/com/back/dao/TelephoneRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.back.entities.Telephone;
 
  
-public interface TelephoneRepository  extends JpaRepository<Telephone, Integer>{
+public interface TelephoneRepository  extends JpaRepository<Telephone, Long>{
 
 }
