+++ src/main/java/com/back/dao/VirementRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.back.entities.Virement;
 
 
-public interface VirementRepository  extends JpaRepository<Virement, Integer> {
+public interface VirementRepository  extends JpaRepository<Virement, Long> {
 
 }
